refactoring pwd hash checks to implement a syscall to do this job. TODO: implment the syscall
