+++ src/include/constants.h
@@ -35,7 +35,9 @@ typedef enum _rm_path_op_t { PROTECT_PATH = 0, UNPROTECT_PATH = 1 } path_op_t;
 
 // THIS SECTION IS ONLY FOR THE KERNEL SPACE
 #ifdef __KERNEL__
+#ifndef MODNAME
 #define MODNAME "KREMFIP"
+#endif
 #include <linux/kernel.h>
 
 // Default sizes of module's hash table
@@ -79,6 +81,16 @@ printk(KERN_INFO "The size of the hash table is too big. We'll reduce to 32 bits
 
 // define the Initial State of the Monitor
 #define RM_INIT_STATE REC_OFF
+
+// define the default crypto algorithm
+#define RM_CRYPTO_ALGO "sha256"
+
+/**
+ * Define the attribute name for the password hash.
+ * BEWARE: this must be the same, at least, of the directory name in the sysfs.
+ */
+#define RM_PWD_HASH_ATTR_NAME "pwd_hash"	// the name of the attribute
+#define RM_PWD_HASH_PATH "/sys/module/kremfip/pwd_hash/pwd_hash" // the path to the attribute
 
 #else
