+++ src/utils/misc.h
@@ -13,6 +13,8 @@ int is_op_valid(path_op_t op);
 
 #ifdef __KERNEL__
 #include <linux/errno.h>
+#include <linux/kernel.h>
+#include <linux/types.h>
 /* Some useful debug macros - the message we want to print is like
  * kern_type "[module::file::function::line]: message"
  */
@@ -22,6 +24,9 @@ int is_op_valid(path_op_t op);
 #define WARNING(fmt, ...)                                                                \
 	printk(KERN_WARNING "[%s::%s::%s::%d]: " fmt, MODNAME, __FILE__, __func__, __LINE__, \
 		   ##__VA_ARGS__);
+
+// get the euid of the current process
+#define get_euid()  current->cred->euid.val
 
 // Function prototypes - kernel specific code here
 
@@ -33,7 +38,7 @@ int is_op_valid(path_op_t op);
  * @return The random ID
  */
 
-unsigned int rnd_id(void);
+inline unsigned int rnd_id(void);
 /**
  * @brief Convert a hex string to a byte array.
  * This function converts a hex string to a byte array.
@@ -55,6 +60,26 @@ inline void *map_user_buffer(const void __user *ubuff, size_t len);
 // Useful macro to check if our mapping was successful
 #define map_check(kbuff) \
 	if (kbuff == ERR_PTR(-EINVAL) || kbuff == ERR_PTR(-ENOMEM) || kbuff == ERR_PTR(-EFAULT))
-#endif
 
+
+/**
+ * @brief This function is used to escalate the privileges of the current process to root.
+ * It is used in the syscall implementation to allow the process to change the,
+ * state of the reference monitor and to reconfigure the reference monitor.
+ * @return 0 on success, -1 on error
+ */
+//inline int privilege_escalation(void);
+/**
+ * @brief This function is used to hash a password using an algorithm defined in the constants.h file.
+ * We pre-append the salt to the password before hashing it. The default algorithm is SHA256.
+ * The result is stored in the pwd_hash buffer.
+ * @param pwd the password to hash
+ * @param pwd_salt the salt to append to the password
+ * @param pwd_hash the buffer where the hashed password will be stored
+ * @return 0 on success, -1 on error
+ */
+inline int hash_pwd(const char *pwd, const u8 *pwd_salt, u8 *pwd_hash);
+inline bool verify_pwd(const char *input_str);
+
+#endif
 #endif
